registry/requirements.txt
# SkillPM Registry - Python Dependencies
# Author: Bogdan Ticu
# License: MIT

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
aiofiles>=23.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Database migrations
alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Production
gunicorn>=21.2.0

# Optional monitoring
prometheus-client>=0.18.0
